#100 DashboardのBuddy表示をDashBoardBuddyとRiveBuddyに切り出し
