+++ frontend_react/public/assets/animations/buddy.riv
