+++ frontend_react/src/pages/dashboard/ui/dashboard.tsx
@@ -1,7 +1,7 @@
 import { PageHeader } from '@/shared/ui/page/page-header';
 import { PageTitle } from '@/shared/ui/page/page-title';
 import { PageTitleContainer } from '@/shared/ui/page/page-title-container';
-import { Buddy } from './buddy';
+import { DashboardBuddy } from './dashboard-buddy';
 import { ToolLink } from './tool-link';
 
 export const Dashboard = () => {
@@ -12,7 +12,7 @@ export const Dashboard = () => {
           <PageTitle>Dashboard</PageTitle>
         </PageTitleContainer>
 
-        <Buddy />
+        <DashboardBuddy />
 
         <div className="space-y-1.5">
           <h3 className="font-bold">Tools</h3>
