+++ frontend_react/src/shared/lib/random-number.ts
@@ -1,0 +1,3 @@
+export function randInt(min: number, max: number): number {
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
